refactor: move reply reads to comment/api/reply
